Bug 479502 - Hold on to a reference to the entry when NotifyExpired is called. r=vlad
The expiration tracker doesn't hold on to references to imgCacheEntries, so if
we're deleting it from the cache, that'll bring its refcount to zero. We
always assume that, when we're calling a function, there exists a reference
that lasts as long as the function does. This patch just makes that true for
the NotifyExpired case.

--HG--
extra : transplant_source : %00%15%EC%27%F98%B4%2C%19%17%9C%F9%3E%5Cg%7F%10%FE%99%01