Bug 514605 - Possible for concurrent reloads of images to return old/different results. r=vlad,jrmuizel
Only set the load ID when we're sure that the entry we're using is valid/going
to be valid. If we're validating an image over the network, we must only set
its load ID after we've validated it.

--HG--
extra : rebase_source : 150bfa4b102f91d870466ef8968d157a78e931a3